Read the file frame by frame
