+++ index.js
@@ -1,8 +1,9 @@
 var events = require('events');
 var util = require('util');
 var async = require('async');
-
 var NobleDevice = require('noble-device');
+var Upload = require('./Upload');
+var Update = require('./Update');
 
 var LIVE_SERVICE_UUID                       = '39e1fa0084a811e2afba0002a5d5c51b';
 var CALIBRATION_SERVICE_UUID                = '39e1fe0084a811e2afba0002a5d5c51b';
@@ -48,7 +49,7 @@ function FlowerPower(peripheral) {
 	this.uuid = peripheral.uuid;
 	this.name = peripheral.advertisement.localName;
 	var flags = peripheral.advertisement.manufacturerData.readUInt8(0);
-	this.flags={};
+	this.flags = {};
 	this.flags.hasEntry = ((flags & (1<<0)) !== 0);
 	this.flags.hasMoved = ((flags & (1<<1)) !== 0);
 	this.flags.isStarting = ((flags & (1<<2)) !== 0);
@@ -476,139 +477,6 @@ FlowerPower.prototype.getStartupTime = function (callback) {
 	});
 };
 
-function UploadBuffer(buffer) {
-	this.idx = buffer.readUInt16LE(0);
-	this.data = new Buffer(buffer.slice(2));
-	return this;
-}
-
-function Upload(fp, callback) {
-	this.fp = fp;
-	this.buffers = [];
-	this.currentIdx = 0;
-	this.RxStatusEnum = {
-		STANDBY: 0,
-		RECEIVING: 1,
-		ACK: 2,
-		NACK: 3,
-		CANCEL: 4,
-		ERROR: 5
-	};
-	this.TxStatusEnum = {
-		IDLE: 0,
-		TRANSFERING: 1,
-		WAITING_ACK: 2
-	};
-	this.rxStatus = this.RxStatusEnum.STANDBY;
-	this.TxStatus = this.TxStatusEnum.IDLE;
-	this.finishCallback = callback;
-	this.startUpload(function(err){
-		if (err !== null) {
-			this.finishCallback(err, null);
-		}
-	});
-	this.fileLength = null;
-	this.bufferLength = null;
-	this.nbTotalBuffers = null;
-	return this;
-
-}
-
-Upload.prototype.onWaitingAck = function(callback) {
-	var success = true;
-	for (var idx=this.currentIdx; ((idx < this.currentIdx+128) && (idx < this.nbTotalBuffers)); idx++) {
-		if (idx>0){
-			if (!this.buffers.hasOwnProperty(idx)){
-				success = false;
-				break;
-			}
-		}
-	}
-	if (success === true) {
-		this.currentIdx += 128;
-		if (this.currentIdx >= this.nbTotalBuffers){
-			this.historyFile = Buffer.concat( this.buffers.slice(1), this.fileLength);
-			async.series([
-					this.writeRxStatus.bind(this, this.RxStatusEnum.ACK),
-					this.unnotifyTxStatus.bind(this),
-					this.unnotifyTxBuffer.bind(this),
-					this.writeRxStatus.bind(this, this.RxStatusEnum.STANDBY)]);
-		}
-		else{
-			async.series([
-					this.writeRxStatus.bind(this, this.RxStatusEnum.ACK)
-					]);
-		}
-	}
-	else {
-		this.writeRxStatus(this.RxStatusEnum.NACK, callback);
-	}
-
-};
-
-Upload.prototype.onTxStatusChange = function (data) {
-	this.txStatus = data.readUInt8(0);
-	if(this.txStatus === this.TxStatusEnum.WAITING_ACK) {
-		this.onWaitingAck();
-	}
-	if(this.txStatus === this.TxStatusEnum.IDLE) {
-		if (this.historyFile !== null && typeof this.historyFile !== 'undefined') {
-			this.finishCallback(null, this.historyFile.toString('base64'));
-			return;
-		}
-		else {
-			this.finishCallback(new Error("Transfer failed", null));
-		}
-	}
-};
-
-Upload.prototype.setFileLength = function (fileLength) {
-	this.fileLength = fileLength;
-	this.nbTotalBuffers = Math.ceil(this.fileLength / this.bufferLength)+1;
-};
-
-Upload.prototype.readFirstBuffer = function (buffer) {
-	this.bufferLength = buffer.length;
-	this.setFileLength(buffer.readUInt32LE(0));
-};
-
-Upload.prototype.onTxBufferReceived = function (data) {
-	var buffer = new UploadBuffer(data);
-	this.buffers[buffer.idx] = buffer.data;
-	if (buffer.idx === 0) {
-		this.readFirstBuffer(buffer.data);
-	}
-};
-
-Upload.prototype.notifyTxStatus = function (callback) {
-	this.fp.notifyCharacteristic(UPLOAD_SERVICE_UUID, UPLOAD_TX_STATUS_UUID, true, this.onTxStatusChange.bind(this), callback);
-};
-
-Upload.prototype.notifyTxBuffer = function (callback) {
-	this.fp.notifyCharacteristic(UPLOAD_SERVICE_UUID, UPLOAD_TX_BUFFER_UUID, true, this.onTxBufferReceived.bind(this), callback);
-};
-
-Upload.prototype.unnotifyTxStatus = function (callback) {
-	this.fp.notifyCharacteristic(UPLOAD_SERVICE_UUID, UPLOAD_TX_STATUS_UUID, false, this.onTxStatusChange.bind(this), callback);
-};
-
-Upload.prototype.unnotifyTxBuffer = function (callback) {
-	this.fp.notifyCharacteristic(UPLOAD_SERVICE_UUID, UPLOAD_TX_BUFFER_UUID, false, this.onTxBufferReceived.bind(this), callback);
-};
-
-Upload.prototype.writeRxStatus = function (rxStatus, callback) {
-	var rxStatusBuff = new Buffer(1);
-	rxStatusBuff.writeUInt8(rxStatus, 0);
-	this.fp.writeDataCharacteristic(UPLOAD_SERVICE_UUID, UPLOAD_RX_STATUS_UUID, rxStatusBuff, callback);
-};
-
-Upload.prototype.startUpload = function (callback) {
-	async.series([
-			this.notifyTxStatus.bind(this),
-			this.notifyTxBuffer.bind(this),
-			this.writeRxStatus.bind(this, this.RxStatusEnum.RECEIVING) ]);
-};
-
 FlowerPower.prototype.getHistory = function (startIdx, callback) {
 	this.writeTxStartIdx(startIdx, function(err) {
 		new Upload(this, callback);
@@ -621,6 +489,10 @@ FlowerPower.prototype.ledPulse = function(callback) {
 
 FlowerPower.prototype.ledOff = function(callback) {
 	this.writeDataCharacteristic(LIVE_SERVICE_UUID, LED_UUID, new Buffer([0x00]), callback);
+};
+
+FlowerPower.prototype.updateFirmware = function(binaryFile, callback) {
+	var update = new Update(this, binaryFile, callback);
 };
 
 module.exports = FlowerPower;
